requirements.txt
flask==3.0.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
scikit-learn==1.3.2
scikit-image==0.22.0
python-dotenv==1.0.0
joblib==1.3.2
